implemented lightning cli to clean things up. Now it's mega easy to run/adjust
