+++ my_config.yaml
@@ -1,0 +1,33 @@
+seed_everything: 42
+
+model:
+  class_weights: [0.1170, 1.4663, 0.6189, 4.9267, 0.8735, 3.4213, 0.7119, 2.7992, 1.5020, 2.2394, 0.7203, 8.2111, 1.1730, 15.3958, 20.5278, 15.3958, 17.5952, 24.6333]
+data:
+  train_path: "trainData.json"
+  val_path: "valData.json"
+  test_path: "testData.json"
+
+trainer:
+  max_epochs: 100
+  devices: 1
+  accelerator: gpu
+  callbacks:
+    - class_path: lightning.pytorch.callbacks.EarlyStopping
+      init_args:
+        monitor: val_f1_micro
+        mode: max
+    - class_path: lightning.pytorch.callbacks.ModelCheckpoint
+      init_args:
+        monitor: val_f1_micro
+        save_top_k: 1
+        mode: max
+        dirpath: "checkpoints/"
+        filename: "best-checkpoint"
+        save_weights_only: true
+  logger:
+    class_path: lightning.pytorch.loggers.WandbLogger
+    init_args:
+      project: GutBrainIE
+      name: cli_test
+      log_model: true
+
