+++ requirements.txt
@@ -2,7 +2,9 @@ aiohappyeyeballs==2.5.0
 aiohttp==3.11.13
 aiosignal==1.3.2
 annotated-types==0.7.0
+antlr4-python3-runtime==4.9.3
 attrs==25.1.0
+bitsandbytes==0.45.5
 blis==0.7.11
 catalogue==2.0.10
 certifi==2025.1.31
@@ -11,18 +13,26 @@ click==8.1.8
 cloudpathlib==0.16.0
 confection==0.1.5
 conllu==6.0.0
+contourpy==1.3.1
+cycler==0.12.1
 cymem==2.0.11
 docker-pycreds==0.4.0
+docstring_parser==0.16
 en_core_sci_sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz#sha256=a54a114ee11dab7cf6db8cf69ef485145011537f45f3c16c71338c48fd6f070b
 filelock==3.17.0
+fonttools==4.57.0
 frozenlist==1.5.0
 fsspec==2025.3.0
 gitdb==4.0.12
 GitPython==3.1.44
 huggingface-hub==0.29.2
+hydra-core==1.3.2
 idna==3.10
+importlib_resources==6.5.2
 Jinja2==3.1.6
 joblib==1.4.2
+jsonargparse==4.38.0
+kiwisolver==1.4.8
 langcodes==3.5.0
 language_data==1.3.0
 lightning==2.5.0.post0
@@ -30,6 +40,7 @@ lightning-utilities==0.14.0
 marisa-trie==1.2.1
 markdown-it-py==3.0.0
 MarkupSafe==3.0.2
+matplotlib==3.8.4
 mdurl==0.1.2
 mpmath==1.3.0
 multidict==6.1.0
@@ -50,7 +61,9 @@ nvidia-cusparselt-cu12==0.6.2
 nvidia-nccl-cu12==2.21.5
 nvidia-nvjitlink-cu12==12.4.127
 nvidia-nvtx-cu12==12.4.127
+omegaconf==2.3.0
 packaging==24.2
+pillow==11.1.0
 platformdirs==4.3.7
 preshed==3.0.9
 propcache==0.3.0
@@ -60,7 +73,9 @@ pybind11==2.13.6
 pydantic==2.10.6
 pydantic_core==2.27.2
 Pygments==2.19.1
+pyparsing==3.2.3
 pysbd==0.3.4
+python-dateutil==2.9.0.post0
 pytorch-lightning==2.5.0.post0
 PyYAML==6.0.2
 regex==2024.11.6
@@ -82,6 +97,7 @@ spacy-legacy==3.0.12
 spacy-loggers==1.0.5
 srsly==2.5.1
 sympy==1.13.1
+tensorboardX==2.6.2.2
 thinc==8.2.5
 threadpoolctl==3.6.0
 tokenizers==0.21.0
@@ -92,6 +108,7 @@ tqdm==4.67.1
 transformers==4.49.0
 triton==3.2.0
 typer==0.9.4
+typeshed_client==2.7.0
 typing_extensions==4.12.2
 urllib3==2.3.0
 wandb==0.19.8
@@ -99,5 +116,3 @@ wasabi==1.1.3
 weasel==0.3.4
 wrapt==1.17.2
 yarl==1.18.3
-matplotlib==3.8.4
-
